+++ apps/cow-fi/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [2.13.0](https://github.com/cowprotocol/cowswap/compare/cow-fi-v2.12.0...cow-fi-v2.13.0) (2026-09-03)
+
+### Minor Changes
+
+- feat: CSS reset additions (button, input) (#8001)
+
+- feat(cow-fi): add get in touch CTA to protocol hero (#8042)
+
+- feat(disclaimer): add content disclaimer component and integrate into… (#8040)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/core@3.8.0
+  - @cowprotocol/common-hooks@3.10.0
+  - @cowprotocol/ui@3.12.0
+  - @cowprotocol/common-const@3.10.0
+  - @cowprotocol/analytics@3.10.0
+  - @cowprotocol/common-utils@3.14.0
+  - @cowprotocol/events@4.11.0
+  - @cowprotocol/types@4.9.0
+  - @cowprotocol/wallet@3.12.0
+  - @cowprotocol/widget-react@3.2.6
+
 ## [2.12.0](https://github.com/cowprotocol/cowswap/compare/cow-fi-v2.11.0...cow-fi-v2.12.0) (2026-08-27)
 
 ### Minor Changes